Make 'half_move_clock' a 'u32'
It *could* be set to a high value due to e.g: starting the engine in the
middle of a game.

Moving from a `u8` to a `u32` does not change the size of the type, so
let's just do that.

Use that opportunity to fix the comment about the number of
*half-moves* (it's 50 moves *per player*).

+++ src/board/chess_board/builder.rs
@@ -8,7 +8,7 @@ pub struct ChessBoardBuilder {
     // Same fields as [ChessBoard].
     castle_rights: [CastleRights; Color::NUM_VARIANTS],
     en_passant: Option<Square>,
-    half_move_clock: u8,
+    half_move_clock: u32,
     side: Color,
     // 1-based, a turn is *two* half-moves (i.e: both players have played).
     turn_count: u32,
@@ -41,7 +41,7 @@ impl ChessBoardBuilder {
         self
     }
 
-    pub fn with_half_move_clock(&mut self, clock: u8) -> &mut Self {
+    pub fn with_half_move_clock(&mut self, clock: u32) -> &mut Self {
         self.half_move_clock = clock;
         self
     }
